Add support for pushing commitmeta objects via garage-push
Push the commitmeta object associated to a commit to TreeHub; this is
done when such object is found in the local repository. The commitmeta
objects can be used to store detached metadata and are particularly
important when composefs is in use by OSTree in which case the commits
would have rootfs signature information kept as detached metadata.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>