Replace Ubuntu Bionic tests with Debian Stable
Ubuntu Bionic is ancient, out of support and ships with a version of boost
that is incompatible with the latest Boost because it doesn't include
boost::asio::io_context, which is the replacement for io_service.

We are already testing PKCS#11 with oldstable, so disabling it on stable
is safe. When someone fixes the PKCS#11 tests for stable, we can re-enable
those tests, but for now the oldstable tests will keep the code working.

Signed-off-by: Phil Wise <phil@phil-wise.com>
Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>