Revert "Set an upper bound of allowed root rotations to 1000."
This reverts commit ff053f9820611c7a54ca5baadccef61d35445c4d.

The upper bound on the number of rotations provides no real protection
against attacks and adds a constraint that will at some point affect
users who frequently rotate their keys.

For Toradex, the key rotation may be triggered when a Lockbox is created
which, for some customers, is a common operation. In fact, we already
have customers facing that issue, which once happens it is hard or even
impossible to fix as the devices can no longer be updated after the
rotations limit is reached. With the present commit, we should prevent
that from happening with other customers.

Related-to: TOR-3850

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>