Ensure Docker tarball is always read to the end
The Docker tarball loading involves reading the tarball twice: on the
first pass the reading is driven by libarchive while the second does not
involve that library. In some cases libarchive does not read the file
until the end because apparently it determines the end of the file has
only padding. When that happens, the validation code was detecting a
size mismatch between the two passes and stopping the tarball loading,
effectively preventing the load of a Lockbox.

The current commit fixes the situation by ensuring that the file is
always read until the end even when the reading is driven by libarchive.

Related-to: TOR-3854

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>