sotauptaneclient: Add condition to handle OSTree hash divergence
This commit attempts to handle the case where the OS (running on the
primary ECU) is updated outside the control of Aktualizr. Before the
commit, the following sequence of steps demonstrates the issue we're
trying to solve:

1. The device has an update version "x" installed.

2. You update OSTree outside of Aktualizr to version "y". This will
not update Aktualizr’s storage, but Aktualizr will see that this
external update happened and report it in its logs and to the
platform.

3. Have the platform update the device to version "x". Aktualizr
will check its storage and see that the latest version is version
"x" and incorrectly ignore the update, as OSTree is actually on
version "y".

With the changes here, we check if the OS version in storage differs
from the currently deployed OSTree commit and, if so, we assume an
incoming target needs to be installed, regardless of it matching the
information in the database.

Related-to: TOR-3518
Signed-off-by: Leonardo Bornia <leonardo.bornia@toradex.com>