Fix race condition in test_imagerepo_failure
When built in TORIZON mode, Aktualizr's default configuration effectively
contains:

    [uptane]
    update_lock_file = "/run/lock/aktualizr-lock"

This causes multiple parallel tests to fail, although the race window is
somewhat narrow. Remove this race condition by setting the config option
to either disable the update locks (for the C++ tests) or set it to a unique
location (for the Python tests).

Signed-off-by: Phil Wise <phil@phil-wise.com>